update examples based on review
