+++ k6/browser/03.thresholds.js
@@ -1,6 +1,5 @@
 import { browser } from "k6/experimental/browser";
 import { check } from "k6";
-import http from "k6/http";
 
 const BASE_URL = __ENV.BASE_URL || "http://localhost:3333";
 
@@ -15,14 +14,11 @@ export const options = {
       },
     },
   },
-};
-
-export function setup() {
-  let res = http.get(BASE_URL)
-  if (res.status !== 200) {
-    throw new Error(`Got unexpected status code ${res.status} when trying to setup. Exiting.`)
+  thresholds: {
+    browser_web_vital_fcp: ["p(95) < 1000"],
+    browser_web_vital_lcp: ["p(95) < 2000"],
   }
-}
+};
 
 export default async function () {
   const page = browser.newPage();
@@ -44,9 +40,4 @@ export default async function () {
   } finally {
     page.close();
   }
-}
-
-export function teardown() {
-  // TODO: Send notification to Slack
-  console.log("That's all folks!")
 }
