+++ k6/browser/05.custom-metrics.js
@@ -6,25 +6,48 @@ const BASE_URL = __ENV.BASE_URL || "http://localhost:3333";
 
 export const options = {
   scenarios: {
-    ui: {
+    pizzaRecommendations: {
       executor: "shared-iterations",
       options: {
         browser: {
           type: "chromium",
         },
       },
+      exec: 'pizzaRecommendations'
+    },
+    admin: {
+      executor: "shared-iterations",
+      options: {
+        browser: {
+          type: "chromium",
+        },
+      },
+      exec: 'admin'
     },
   },
   thresholds: {
     browser_web_vital_fcp: ["p(95) < 1000"],
     browser_web_vital_lcp: ["p(95) < 2000"],
-    checks: ["rate > 0.9"]
   }
 };
 
 const myTrend = new Trend('totalActionTime');
 
-export default async function () {
+export async function admin() {
+  const page = browser.newPage();
+
+  try {
+    await page.goto(`${BASE_URL}/admin`);
+    await page.locator('button[type="submit"]').click();
+    check(page, {
+      "logout button text": page.locator('//*[text()="Logout"]').textContent() == "Logout",
+    });
+  } finally {
+    page.close();
+  }
+}
+
+export async function pizzaRecommendations() {
   const page = browser.newPage();
 
   try {
@@ -39,9 +62,9 @@ export default async function () {
 
     await page.locator('//button[. = "Pizza, Please!"]').click();
     page.waitForTimeout(500);
+    page.screenshot({ path: "screenshot.png" });
     page.evaluate(() => window.performance.mark('recommendations-returned'));
 
-    page.screenshot({ path: "screenshot.png" });
     check(page, {
       recommendation: page.locator("div#recommendations").textContent() != "",
     });
