+++ k6/foundations/08.arrival-rate.js
@@ -25,6 +25,7 @@ export const options = {
     http_req_failed: ['rate<0.01'],
     http_req_duration: ['p(95)<500', 'p(99)<1000'],
     quickpizza_ingredients: [{ threshold: 'avg<8', abortOnFail: false }],
+    checks: ["rate > 0.95"]
   },
 };
 
@@ -73,7 +74,7 @@ export function getPizza() {
   */
 }
 
-export function teardown(){
+export function teardown() {
   // TODO: Send notification to Slack
   console.log("That's all folks!")
 }
