+++ k6/foundations/10.summary.js
@@ -2,54 +2,25 @@ import http from "k6/http";
 import { check, sleep } from "k6";
 import { Trend, Counter } from "k6/metrics";
 import { textSummary } from "https://jslib.k6.io/k6-summary/0.0.2/index.js";
-import { SharedArray } from 'k6/data';
-import { browser } from "k6/experimental/browser";
 
 const BASE_URL = __ENV.BASE_URL || 'http://localhost:3333';
 
 export const options = {
-  scenarios: {
-    smoke: {
-      exec: "getPizza",
-      executor: "constant-vus",
-      vus: 1,
-      duration: "10s",
-    },
-    stress: {
-      exec: "getPizza",
-      executor: "ramping-vus",
-      stages: [
-        { duration: '5s', target: 5 },
-        { duration: '10s', target: 5 },
-        { duration: '5s', target: 0 },
-      ],
-      startTime: "10s",
-    },
-    browser: {
-      exec: "checkFrontend",
-      executor: "constant-vus",
-      vus: 1,
-      duration: "30s",
-        options: {
-          browser: {
-            type: "chromium",
-          },
-      },
-    }
-  },
+  stages: [
+    { duration: '5s', target: 5 },
+    { duration: '10s', target: 5 },
+    { duration: '5s', target: 0 },
+  ],
   thresholds: {
     http_req_failed: ['rate<0.01'],
     http_req_duration: ['p(95)<500', 'p(99)<1000'],
     quickpizza_ingredients: [{ threshold: 'avg<8', abortOnFail: false }],
+    checks: ["rate > 0.95"]
   },
 };
 
 const pizzas = new Counter('quickpizza_number_of_pizzas');
 const ingredients = new Trend('quickpizza_ingredients');
-
-const customers = new SharedArray('all my customers', function () {
-  return JSON.parse(open('./data/customers.json')).customers;
-});
 
 export function setup() {
   let res = http.get(BASE_URL)
@@ -58,7 +29,7 @@ export function setup() {
   }
 }
 
-export function getPizza() {
+export default function () {
   let restrictions = {
     maxCaloriesPerSlice: 500,
     mustBeVegetarian: false,
@@ -70,7 +41,7 @@ export function getPizza() {
   let res = http.post(`${BASE_URL}/api/pizza`, JSON.stringify(restrictions), {
     headers: {
       'Content-Type': 'application/json',
-      'X-User-ID': customers[Math.floor(Math.random() * customers.length)],
+      'X-User-ID': 23423,
     },
   });
   check(res, { "status is 200": (res) => res.status === 200 });
@@ -80,27 +51,7 @@ export function getPizza() {
   sleep(1);
 }
 
-export async function checkFrontend() {
-  const page = browser.newPage();
-
-  try {
-    await page.goto(BASE_URL)
-    check(page, {
-      'header': page.locator('h1').textContent() == 'Looking to break out of your pizza routine?',
-    });
-
-    await page.locator('//button[. = "Pizza, Please!"]').click();
-    page.waitForTimeout(500);
-    page.screenshot({ path: `screenshots/${__ITER}.png` });
-    check(page, {
-      'recommendation': page.locator('div#recommendations').textContent() != '',
-    });
-  } finally {
-    page.close();
-  }
-}
-
-export function teardown(){
+export function teardown() {
   // TODO: Send notification to Slack
   console.log("That's all folks!")
 }
